BACKEND/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.4
werkzeug==2.2.2
gunicorn==20.1.0
python-dotenv==0.21.0
google-generativeai==0.2.2
google-auth==2.16.0
torch
torchvision
Pillow